python/sample-app-pit/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.0.0
python-dotenv>=0.19.0
gohighlevel-api-client==1.0.0b1